requirements_analysis.txt
aiohttp==3.7.4
allensdk==2.13.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argschema==3.0.4
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bg-atlasapi==1.0.2
bg-space==0.6.0
bleach==5.0.1
boto3==1.17.21
botocore==1.20.112
brainrender==2.0.5.3
bs4==0.0.1
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distro==1.7.0
docxcompose==1.3.5
docxtpl==0.16.3
entrypoints==0.4
fastjsonschema==2.15.3
future==0.18.2
Glymur==0.8.19
google==3.0.0
greenlet==1.1.2
h5py==3.7.0
hdmf==3.3.2
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipydatawidgets==4.3.1.post1
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
K3D==2.7.4
kiwisolver==1.4.3
littleutils==0.2.2
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.0.1
marshmallow==3.17.0
matplotlib==3.4.2
matplotlib-inline==0.1.3
meshio==5.3.4
mistune==0.8.4
morphapi==0.1.7
MorphIO==3.3.3
msgpack==1.0.4
multidict==6.0.2
myterial==1.2.1
natsort==8.1.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.4.0
ndx-events==0.2.0
nest-asyncio==1.5.5
networkx==2.6.3
neurom==3.2.0
notebook==6.4.12
numexpr==2.8.3
numpy==1.21.6
outdated==0.2.1
packaging==21.3
pandas==1.3.5
pandas-flavor==0.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pingouin==0.5.2
pip==22.1.2
pipenv==2018.11.26
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.18.1
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.12.0
pyinspect==0.1.0
pylustrator==1.1.2
pynrrd==0.4.3
pynwb==2.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
QtAwesome==1.1.1
QtPy==2.1.0
requests==2.28.1
requests-toolbelt==0.9.1
retry==0.9.2
rich==12.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
scikit-build==0.15.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
setuptools==63.1.0
SimpleITK==2.1.1.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
statannotations==0.4.4
statsmodels==0.13.0
tables==3.6.1
tabulate==0.8.10
terminado==0.15.0
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
traittypes==0.2.1
treelib==1.6.1
typing_extensions==4.3.0
urllib3==1.26.10
vedo==2021.0.5
virtualenv==16.7.8
virtualenv-clone==0.5.3
vtk==8.1.2
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.6.1
wrapt==1.14.1
wslink==1.6.6
xarray==0.20.2
yarl==1.7.2
zipp==3.8.1
